fix dead imports and move metadata-related objects into metadata.py
