+++ bin/git-theta-filter
@@ -4,7 +4,7 @@ import argparse
 import sys
 import logging
 
-from git_theta import git_utils, checkpoints, params, updates
+from git_theta import git_utils, checkpoints, params, metadata, updates
 
 logging.basicConfig(
     level=logging.DEBUG,
@@ -60,17 +60,17 @@ def clean(args):
 
     obj = git_utils.get_file_version(repo, args.file, "HEAD")
     if obj:
-        metadata = params.Metadata.from_file(obj.data_stream).flatten()
+        curr_metadata = metadata.Metadata.from_file(obj.data_stream).flatten()
     else:
-        metadata = params.Metadata()
+        curr_metadata = metadata.Metadata()
 
-    new_metadata = params.Metadata()
+    new_metadata = metadata.Metadata()
     # Sort the keys so we don't get changing diffs based on serialization order.
     for param_keys, new_param in sorted(model_checkpoint.flatten().items()):
         logging.debug(f"Cleaning {'/'.join(param_keys)}")
 
-        param_metadata = metadata.get(param_keys)
-        new_tensor_metadata = params.TensorMetadata.from_tensor(new_param)
+        param_metadata = curr_metadata.get(param_keys)
+        new_tensor_metadata = metadata.TensorMetadata.from_tensor(new_param)
         # If the parameter tensor has not changed, just keep the metadata the same
         #TODO: currently checks if rounded parameter hash has changed. Look into better parameter equality checks.
         if param_metadata and param_metadata.tensor_metadata == new_tensor_metadata:
@@ -80,11 +80,11 @@ def clean(args):
             update = updates.get_update()()
             update_serializer = params.get_update_serializer()
 
-            new_theta_metadata = params.ThetaMetadata(update_type=update.name, last_commit=git_utils.get_head(repo))
+            new_theta_metadata = metadata.ThetaMetadata(update_type=update.name, last_commit=git_utils.get_head(repo))
             param_update = update.calculate_update(repo, args.file, param_keys, param_metadata, new_param)
             serialized_update = update_serializer.serialize(param_update)
-            new_lfs_metadata = params.LfsMetadata.from_bytes(serialized_update)
-            new_param_metadata = params.ParamMetadata(lfs_metadata=new_lfs_metadata,
+            new_lfs_metadata = metadata.LfsMetadata.from_bytes(serialized_update)
+            new_param_metadata = metadata.ParamMetadata(lfs_metadata=new_lfs_metadata,
                                                       tensor_metadata=new_tensor_metadata,
                                                       theta_metadata=new_theta_metadata)
 
@@ -99,10 +99,10 @@ def smudge(args):
     logging.debug(f"Running smudge filter on {args.file}")
 
     repo = git_utils.get_git_repo()
-    metadata = params.Metadata.from_file(sys.stdin).flatten()
+    curr_metadata = metadata.Metadata.from_file(sys.stdin).flatten()
 
     model_dict = {}
-    for param_keys, param_metadata in metadata.items():
+    for param_keys, param_metadata in curr_metadata.items():
         logging.debug(f"Smudging {'/'.join(param_keys)}")
         update = updates.get_update(param_metadata.theta_metadata.update_type)()
         model_dict[param_keys] = update.apply(repo, args.file, param_keys, param_metadata)
